 requirements.txt
flask>=2.0.0
tensorflow>=2.13.0
pillow>=9.0.0
numpy>=1.22.0
keras>=3.8.0
onnx>=1.14.0
onnxruntime>=1.16.0